requirements.txt
aiohttp==2.3.9
SQLAlchemy==1.2
psycopg2==2.7.3
sqlalchemy-mptt==0.2.2
lxml
factory_boy
requests
passlib==1.7.1
pyjwt==1.5.3